requirements.txt
# Core Django
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-ratelimit==4.1.0
django-filter==24.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Caching & Redis
django-redis==5.4.0
redis==5.0.1

# Payments
razorpay==2.0.0

# Email
django-anymail==10.2
resend==2.19.0

# File Storage
django-storages==1.14.2
boto3==1.34.67
cloudinary==1.36.0
django-cloudinary-storage==0.3.0

# API & Serialization
requests==2.32.3
python-dateutil==2.8.2

# Video Embedding
django-embed-video==1.4.10

# Development & Testing
django-extensions==3.2.3
factory-boy==3.3.0
faker==22.6.0

# Static Files
whitenoise==6.6.0

# Web Server
gunicorn==22.0.0

# Error Tracking (optional)
sentry-sdk==2.49.0

# Utilities
python-decouple==3.8
python-dotenv==1.0.0
pytz==2024.1
